fixes failing tests for AssertionMacro
In ruby 2.4 a change was made to allow refinement of certain methods
like send. The Riot::Message::BlankSlate class was subsequently
removing the #send and #kind_of? method (as well as others). Previously
it was avoiding the removal of any methods starting with "__". The change
here is to actually fix the test to not use Message as the topic for the
failing tests.